refactor: reorganize customer reservation components
