+++ apps/ui/components/reservation/styles.ts
@@ -28,3 +28,20 @@ export const PinkBox = styled.div`
   padding: var(--spacing-m);
   background-color: var(--color-suomenlinna-light);
 `;
+
+export const PreviewLabel = styled.div`
+  color: var(--color-black-70);
+  padding-bottom: var(--spacing-2-xs);
+`;
+
+export const PreviewValue = styled.div`
+  font-size: var(--fontsize-body-l);
+`;
+
+export const ParagraphAlt = styled.div<{ $isWide?: boolean }>`
+  ${({ $isWide }) => $isWide && "grid-column: 1 / -1;"}
+
+  & > div:first-of-type {
+    margin-bottom: var(--spacing-3-xs);
+  }
+`;
